remove timeit calls, make late tag show up in detail view, move find_next_leaveslip to utils file, fix dumb mistake
